Move util to different file
